Refactor to remove duplicate code, and remove unused variables For GPX, we now scan for lures and use all additional logic
